update issuer help
