+++ app/components/Help/index.js
@@ -1,3 +1,3 @@
-export * from './IssuerHelp';
+export * from './IssuerHelp/IssuerHelp';
 export * from './ExpirationDateHelp/ExpirationDateHelp';
 export * from './RegistryHelp';
